masih belum keluar output
